+++ CHANGELOG.md
@@ -5,6 +5,11 @@ is based on https://keepachangelog.com/en/1.1.0/.
 
 ## Unreleased
 
+### Added
+
+- `prof` subcommand for generating a profile in the callgrind format from a
+  trace file.
+
 ## 0.1.0 - 2026-03-19
 
 ### Added